+++ decorators/default.decorator.ts
@@ -1,0 +1,5 @@
+import { Transform } from 'class-transformer';
+
+export function Default(defaultValue: unknown): PropertyDecorator {
+	return Transform((value: unknown) => value ?? JSON.parse(JSON.stringify(defaultValue)));
+}