+++ factories/entity.factory.ts
@@ -1,8 +1,8 @@
 import { plainToClassFromExist } from 'class-transformer';
 
 export class EntityFactory {
-  public static create<T>(cls: ClassType<T>, data?: DeepPartial<T>): T {
-    const instance = new cls();
-    return plainToClassFromExist(instance, data || {});
-  }
+	public static create<T>(cls: ClassType<T>, data?: DeepPartial<T>): T {
+		const instance = new cls();
+		return plainToClassFromExist(instance, data || {});
+	}
 }